move type set in another file
