+++ core/src/test/scala/org/apache/spark/sql/IssueTestSuite.scala
@@ -15,80 +15,10 @@
 
 package org.apache.spark.sql
 
-import com.pingcap.tikv.types.Converter
 import com.pingcap.tispark.TiConfigConst
 import org.apache.spark.sql.functions.{col, sum}
 
 class IssueTestSuite extends BaseTiSparkSuite {
-  test("adding time type index test") {
-    tidbStmt.execute("drop table if exists t_t")
-    tidbStmt.execute("CREATE TABLE `t_t` (`t` time(3), index `idx_t`(t))")
-    // NOTE: jdbc only allows time in day range whereas mysql time has much
-    // larger range.
-    tidbStmt.execute("INSERT INTO t_t (t) VALUES('18:59:59'),('17:59:59'),('12:59:59')")
-    refreshConnections()
-    val df = spark.sql("select * from t_t")
-    val data = dfData(df, df.schema.fields)
-    assert(data(0)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("18:59:59")))
-    assert(data(1)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("17:59:59")))
-    assert(data(2)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("12:59:59")))
-
-    val where = spark.sql("select * from t_t where t = 46799000000000")
-    val wheredata = dfData(where, where.schema.fields)
-    assert(wheredata(0)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("12:59:59")))
-  }
-
-  test("adding time type") {
-    tidbStmt.execute("drop table if exists t_t")
-    tidbStmt.execute("CREATE TABLE `t_t` (`t` time(3))")
-    // NOTE: jdbc only allows time in day range whereas mysql time has much
-    // larger range.
-    tidbStmt.execute("INSERT INTO t_t (t) VALUES('18:59:59'),('17:59:59'),('12:59:59')")
-    refreshConnections()
-    val df = spark.sql("select * from t_t")
-    val data = dfData(df, df.schema.fields)
-    assert(data(0)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("18:59:59")))
-    assert(data(1)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("17:59:59")))
-    assert(data(2)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("12:59:59")))
-
-    val where = spark.sql("select * from t_t where t = 46799000000000")
-    val wheredata = dfData(where, where.schema.fields)
-    assert(wheredata(0)(0).asInstanceOf[Long].equals(Converter.convertStrToDuration("12:59:59")))
-  }
-
-  test("adding year type") {
-    tidbStmt.execute("drop table if exists y_t")
-    tidbStmt.execute("CREATE TABLE `y_t` (`y4` year(4))")
-    tidbStmt.execute("INSERT INTO y_t (y4) VALUES(1912),(2012),(2112)")
-    refreshConnections()
-    judge("select * from y_t")
-    judge("select * from y_t where y4 = 2112")
-  }
-
-  test("adding set and enum") {
-    tidbStmt.execute("drop table if exists set_t")
-    tidbStmt.execute("drop table if exists enum_t")
-    tidbStmt.execute(
-      "CREATE TABLE `set_t` (" +
-        "`priority` set('Low','Medium','High') NOT NULL)"
-    )
-    tidbStmt.execute("INSERT INTO set_t(priority) VALUES('High')")
-    tidbStmt.execute("INSERT INTO set_t(priority) VALUES('Medium')")
-    tidbStmt.execute("INSERT INTO set_t(priority) VALUES('Low')")
-    tidbStmt.execute(
-      "CREATE TABLE `enum_t` (" +
-        "`priority` set('Low','Medium','High') NOT NULL)"
-    )
-    tidbStmt.execute("INSERT INTO enum_t(priority) VALUES('High')")
-    tidbStmt.execute("INSERT INTO enum_t(priority) VALUES('Medium')")
-    tidbStmt.execute("INSERT INTO enum_t(priority) VALUES('Low')")
-    refreshConnections()
-    judge("select * from set_t")
-    judge("select * from set_t where priority = 'High'")
-    judge("select * from enum_t")
-    judge("select * from enum_t where priority = 'High'")
-  }
-
   test("cannot resolve column name when specifying table.column") {
     spark.sql("select full_data_type_table.id_dt from full_data_type_table").explain(true)
     judge("select full_data_type_table.id_dt from full_data_type_table")
@@ -100,35 +30,6 @@ class IssueTestSuite extends BaseTiSparkSuite {
     judge(
       "select full_data_type_table.id_dt from full_data_type_table join full_data_type_table_idx on full_data_type_table.id_dt = full_data_type_table_idx.id_dt"
     )
-  }
-
-  test("partition read") {
-    tidbStmt.execute("DROP TABLE IF EXISTS `partition_t`")
-    tidbStmt.execute("""
-                       |CREATE TABLE `partition_t` (
-                       |  `id` int(11) DEFAULT NULL,
-                       |  `name` varchar(50) DEFAULT NULL,
-                       |  `purchased` date DEFAULT NULL
-                       |) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin
-                       |PARTITION BY RANGE ( `id` ) (
-                       |  PARTITION p0 VALUES LESS THAN (1990),
-                       |  PARTITION p1 VALUES LESS THAN (1995),
-                       |  PARTITION p2 VALUES LESS THAN (2000),
-                       |  PARTITION p3 VALUES LESS THAN (2005)
-                       |)
-      """.stripMargin)
-    tidbStmt.execute("insert into partition_t values (1, \"dede1\", \"1989-01-01\")")
-    tidbStmt.execute("insert into partition_t values (2, \"dede2\", \"1991-01-01\")")
-    tidbStmt.execute("insert into partition_t values (3, \"dede3\", \"1996-01-01\")")
-    tidbStmt.execute("insert into partition_t values (4, \"dede4\", \"1998-01-01\")")
-    tidbStmt.execute("insert into partition_t values (5, \"dede5\", \"2001-01-01\")")
-    tidbStmt.execute("insert into partition_t values (6, \"dede6\", \"2006-01-01\")")
-    tidbStmt.execute("insert into partition_t values (7, \"dede7\", \"2007-01-01\")")
-    tidbStmt.execute("insert into partition_t values (8, \"dede8\", \"2008-01-01\")")
-    refreshConnections()
-    assert(spark.sql("select * from partition_t").count() == 8)
-    judge("select count(*) from partition_t where id = 1", checkLimit = false)
-    judge("select id from partition_t group by id", checkLimit = false)
   }
 
   test("test date") {
